Add files via upload
In C++, a class is a blueprint for creating objects that encapsulate data and functions. It consists of data members (attributes) and member functions (methods). The class is defined using the class keyword, and it has three main access specifiers: private, protected, and public. Data in the private section is only accessible within the class, while the public section provides an interface to access the class's functionalities from outside. Classes support object-oriented principles like encapsulation, inheritance, and polymorphism, making C++ powerful for building modular and reusable code.